Auto import message libraries and split code into files according to functionality
